+++ src/dsg/error.rs
@@ -1,6 +1,5 @@
 //! Data Security Guard specific Errors
 
-use actix_web::ResponseError;
 use derive_more::Display;
 use std::error;
 
@@ -36,8 +35,6 @@ pub enum Error {
 }
 
 impl error::Error for Error {}
-
-impl ResponseError for Error {}
 
 #[cfg(test)]
 mod tests {
