+++ src/dtc/error.rs
@@ -1,6 +1,5 @@
 //! Data Tracker Chain specific Errors
 
-use actix_web::ResponseError;
 use derive_more::Display;
 use std::error;
 
@@ -24,8 +23,6 @@ pub enum Error {
 }
 
 impl error::Error for Error {}
-
-impl ResponseError for Error {}
 
 #[cfg(test)]
 mod tests {
