+++ src/dtc/mod.rs
@@ -437,8 +437,6 @@ impl Tracker {
 }
 
 pub mod error;
-pub mod extractor;
-pub mod middleware;
 
 // Unit Tests
 #[cfg(test)]
