--- src/dua/extractor/mod.rs
@@ -1,8 +1,0 @@
-//! An Extractor that parses the HTTP header and pulls out a list of DUAs
-
-use super::*;
-extern crate actix_service;
-extern crate actix_web;
-extern crate futures;
-
-pub mod actix;
