+++ src/dua/mod.rs
@@ -155,8 +155,6 @@ impl DUA {
 }
 
 pub mod error;
-pub mod extractor;
-pub mod middleware;
 
 // Unit Tests
 #[cfg(test)]
